compat+fix: PASS_MAX undef in shader.h; GDI font/DC stubs; fix ptr cast
- shader.h (both Generals and GeneralsMD): extend POSIX PASS_MAX undef
  from Solaris-only to all non-Windows builds. always.h -> <assert.h> ->
  <sys/syslimits.h> defines PASS_MAX=128 on macOS/Linux, which conflicts
  with ShaderClass::DepthCompareType::PASS_MAX enum member.
- windows.h: add GDI font/text/DC stubs needed by render2dsentence.cpp
  on Linux native build: FW_BOLD/NORMAL/THIN, DEFAULT_CHARSET,
  OUT_DEFAULT_PRECIS, CLIP_DEFAULT_PRECIS, ANTIALIASED_QUALITY,
  VARIABLE_PITCH, ETO_OPAQUE, LOGFONT, TEXTMETRIC, MulDiv, CreateFont,
  CreateCompatibleDC, SelectObject, DeleteDC, SetBkColor, SetTextColor,
  GetTextMetrics, ExtTextOutW, GetTextExtentPoint32W.
- assetmgr.cpp: fix ptr-to-int cast: ((int)ptr1) - ((int)ptr2) ->
  (int)(ptr1 - ptr2) using ptrdiff via pointer subtraction.

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WW3D2/assetmgr.cpp
@@ -810,7 +810,7 @@ RenderObjClass * WW3DAssetManager::Create_Render_Obj(const char * name)
 		char filename [MAX_PATH];
 		const char *mesh_name = ::strchr (name, '.');
 		if (mesh_name != NULL) {
-			::lstrcpyn (filename, name, ((int)mesh_name) - ((int)name) + 1);
+			::lstrcpyn (filename, name, (int)(mesh_name - name) + 1);
 			::lstrcat (filename, ".w3d");
 		} else {
 			sprintf( filename, "%s.w3d", name);
